+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ## Unreleased/Future
 
+- Parent methods now returns `None` if `z<=0`
 - New command(s):
   - `agg-hash` command that computes the `agg-tiles-hash` of a tiles-db as standardized by the martin/maplibre team (this supports more hash-types; `xxh3` appears to be the fastest and what utiles will likely default to if not `xxh64`)
   - `commands` list all available commands (including hidden/dev/unimplemented commands)